fix(d1.2): build-green — full workspace compiles in CI Docker (ARM64)
All 15 ROS2 packages now compile clean in mass-l3/ci:jazzy-ubuntu22.04
(ARM64) with -Wall -Wextra -Wpedantic -Werror.

Key fixes across M1-M8, fcb_simulator, integration tests, and Docker:

M1 ODD Envelope Manager
- CMakeLists.txt: gtest ABI workaround (recompile from vendor source)
- Parameter loading: static_cast narrowing conversions (int64_t→int32_t)
- Node headers: nodiscard + missing includes

M2 World Model
- CMakeLists.txt: gtest from vendor source, custom m2_add_gtest macro
- track_buffer.hpp: remove const from compute_control declaration
- world_state_aggregator.cpp: suppress -Werror=float-equal with epsilon check
- Test files: IDL field name fixes (r_dot_deg_s, nav_mode, fill() vs assign())

M3 Mission Manager
- voyage_task_validator: remove const-qualified method mutations
- eta_projector: static_cast for sign-conversion

M5 Tactical Planner
- CMakeLists.txt: CasADi ABI patch (strip _GLIBCXX_USE_CXX11_ABI=0 from
  INTERFACE_COMPILE_DEFINITIONS — pip-installed casadi-config.cmake forces
  old ABI but libcasadi.so.3.7 uses new ABI)
- BC-MPC tests: BcMpcBranchFormulation requires explicit Config{} construction
- Mid-MPC tests: static_cast<void>(solver.solve(...)) for nodiscard
- gtest macro: add --gtest_output=xml for colcon test-result parsing
- pybind11: FindPython3 with Development.Module before pybind11_vendor

M6 COLREGs Reasoner
- CMakeLists.txt: add -Wshadow -Wconversion flags, fix test/CMakeLists.txt
- Rule implementations: fix various -Werror=conversion issues
- colregs_reasoner_node.cpp: nodiscard timer/subscription returns

M7 Safety Supervisor
- Multiple nodiscard fixes (trigger, tick, create_*)
- sha256.cpp: cast alignment, static_cast for narrowing

M8 HMI Transparency Bridge
- on_acknowledgment_clicked: static_cast<void>() on nodiscard return
- trigger/tick: nodiscard fixes

fcb_simulator
- compute_control: remove const (mutates override_active_)
- CMakeLists.txt: FindPython3 with Development.Module

l3_integration_tests
- CMakeLists.txt: CMAKE_MODULE_PATH for FindGeographicLib.cmake
- All 7 test files: OddState→ODDState, eta_window field names,
  tor_request include path

Docker / toolchain
- Dockerfile.ci: pin CasADi 3.7.2, add pybind11_vendor, IPOPT runtime libs
- version-lock.txt: record pinned versions

+++ src/m5_tactical_planner/src/mid_mpc/mid_mpc_nlp_formulation.cpp
@@ -297,7 +297,7 @@ MidMpcSolution MidMpcNlpFormulation::unpack_solution(
   }
   if (stats.count("iter_count") > 0u) {
     sol.ipopt_iterations = static_cast<int32_t>(
-        static_cast<int64_t>(stats.at("iter_count")));
+        static_cast<int>(stats.at("iter_count")));
   }
   return sol;
 }
